refactor: 화폐 단위 자료형 변경(int/double -> BigDecimal) 및 제약 사항 추가

+++ src/main/java/domain/dto/ProfitResponse.java
@@ -1,7 +1,9 @@
 package domain.dto;
 
+import java.math.BigDecimal;
+
 public record ProfitResponse(
         String name,
-        int profit
+        BigDecimal profit
 ) {
 }
